Organize url helpers in a single file and rename "mergeQueryStringsWithData" to "mergeQueryStringWithData"
