+++ packages/inertia/src/index.js
@@ -1,5 +1,3 @@
-export { default as hrefToUrl } from './hrefToUrl'
 export { default as Inertia } from './inertia'
-export { default as mergeQueryStringsWithData } from './mergeQueryStringsWithData'
 export { default as shouldIntercept } from './shouldIntercept'
-export { default as urlWithoutHash } from './urlWithoutHash'
+export { hrefToUrl, mergeQueryStringWithData, urlWithoutHash } from './url'
