--- packages/inertia/src/urlWithoutHash.js
@@ -1,5 +1,0 @@
-export default function urlWithoutHash(url) {
-  url = new URL(url.href)
-  url.hash = ''
-  return url
-}
